cv/grading_model.py
import random
import os

# Try importing torch, else fallback to mock
try:
    import torch
    import torch.nn as nn
    import torchvision.transforms as transforms
    from PIL import Image
    TORCH_AVAILABLE = True
except ImportError:
    TORCH_AVAILABLE = False
    print("Warning: PyTorch not found. Using Mock Grading Engine.")

class GradingModel:
    def __init__(self):
        self.classes = ['Grade A', 'Grade B', 'Grade C']
        self.model = None
        if TORCH_AVAILABLE:
            self.model = self._build_simple_cnn()
            # self.model.load_state_dict(torch.load('weights.pth')) # Uncomment when weights exist
            self.model.eval()

    def _build_simple_cnn(self):
        if not TORCH_AVAILABLE: return None
        # Simple CNN Architecture for Demo
        class SimpleCNN(nn.Module):
            def __init__(self):
                super(SimpleCNN, self).__init__()
                self.conv1 = nn.Conv2d(3, 16, 3, 1)
                self.conv2 = nn.Conv2d(16, 32, 3, 1)
                self.fc1 = nn.Linear(32 * 54 * 54, 128) # Assuming 224x224 input -> ~54x54 after pools
                self.fc2 = nn.Linear(128, 3)

            def forward(self, x):
                # Placeholder forward pass
                return x
        return SimpleCNN()

    def preprocess_image(self, image_path):
        if not TORCH_AVAILABLE: return None
        transform = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.ToTensor(),
        ])
        try:
            img = Image.open(image_path)
            return transform(img).unsqueeze(0)
        except Exception as e:
            print(f"Error loading image: {e}")
            return None

    def predict(self, image_path):
        """
        Returns prediction dict: {grade: str, confidence: float}
        """
        # 1. REAL INFERENCE (If model & weights existed)
        # if TORCH_AVAILABLE and self.model:
        #     tensor = self.preprocess_image(image_path)
        #     if tensor is not None:
        #         with torch.no_grad():
        #             output = self.model(tensor)
        #             # Logic to get argmax...
        
        # 2. MOCK INFERENCE (For Demo/Hackathon usage)
        # Since we don't have a trained .pth file yet, we simulate smart detection.
        # In a real demo, we might use colour histograms to hint at quality, 
        # but random is sufficient for logic flow testing.
        
        mock_grade = random.choice(self.classes)
        # Make Grade A slightly rarer for realism
        if random.random() > 0.7:
            mock_grade = 'Grade A'
        elif random.random() > 0.4:
            mock_grade = 'Grade B'
        else:
            mock_grade = 'Grade C'
            
        return {
            "grade": mock_grade,
            "confidence": round(random.uniform(0.75, 0.98), 2),
            "details": "Grading based on surface texture and color consistency."
        }

# Usage
if __name__ == "__main__":
    grader = GradingModel()
    print(grader.predict("test.jpg"))
