requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
xgboost
requests
beautifulsoup4
feedparser
nltk
streamlit-oauth
scipy
Pillow
# Force Cloud Rebuild
